itemSlot
now when player picks up an item it now displays the name of the item and the description of the item. Player can also click on the individual item slots in the inventory and select different slots in the inventory. Next step is to be able to deselect an already selected item.