FirearmChanges
When modifying the firearm script for the bullet to travel to the reticle it created and issue with enemies firing at players reticle. this issue has been resolved as I had to create a enemy bullet script and an enemy firearm script that way the bullets could be instantiated seperatly.